Fixed typo. Moved utils to util folder
